Added water, upgraded terrain generator, delted useless, some improvments

+++ Minecraft/Player.cpp
@@ -13,9 +13,10 @@ using namespace World;
 void Player::init(Map *map)
 {
 	m_map = map;
-	m_speed = 5*COORDS_IN_BLOCK;
-	m_size = { COORDS_IN_BLOCK / 4.F, COORDS_IN_BLOCK, COORDS_IN_BLOCK / 4.F };
-	m_pos = { 320, 51+30, 360 };
+	m_speed = DEFAULT_PLAYER_SPEED;
+	m_size = { 1.f / 4.F, 1.f, 1.f / 4.F };
+	//m_pos = { 320, 51+30, 360 };
+	m_pos = { 5356, 80, 1495 };
 
 	for (block_id type : DB::s_loaded_blocks)
 		m_inventory.emplace_back(type, 1);
@@ -31,7 +32,7 @@ void Player::input(sf::Event& e)
 
 void Player::update(float dtime)
 {
-	static float acceleration = 5.f*m_speed;// DEFAULT_PLAYER_SPEED;
+	static float acceleration = 2.f*m_speed;// DEFAULT_PLAYER_SPEED m_speed;
 	
 	auto get_oveclocking_moving = [&](Key key) -> float {
 		float dmoving = m_speed*dtime;
@@ -79,46 +80,46 @@ void Player::update(float dtime)
 	};
 
 	m_dpos.x = 0.f;
-	if (m_flying)
-		m_dpos.y = 0.f;
-
+	m_dpos.y = 0.f;
 	m_dpos.z = 0.f;
 
 	// forward
-	float dmoving = get_moving(Key::W)*5;
+	float dmoving = get_moving(Key::W)*1.5f;
 	m_dpos.x += -sinf(m_camera_angle.x / 180 * PI) * dmoving;
 	m_dpos.z += -cosf(m_camera_angle.x / 180 * PI) * dmoving;
 
 	// back
-	dmoving = get_moving(Key::S)*5;
+	dmoving = get_moving(Key::S)*1.5f;
 	m_dpos.x += +sinf(m_camera_angle.x / 180 * PI) * dmoving;
 	m_dpos.z += +cosf(m_camera_angle.x / 180 * PI) * dmoving;
 	
 	// left
-	dmoving = get_moving(Key::A)*5;
+	dmoving = get_moving(Key::A)*1.5f;
 	m_dpos.x += +sinf((m_camera_angle.x - 90) / 180 * PI) * dmoving;
 	m_dpos.z += +cosf((m_camera_angle.x - 90) / 180 * PI) * dmoving;
 
 	//rigth
-	dmoving = get_moving(Key::D)*5;
+	dmoving = get_moving(Key::D)*1.5f;
 	m_dpos.x += +sinf((m_camera_angle.x + 90) / 180 * PI) * dmoving;
 	m_dpos.z += +cosf((m_camera_angle.x + 90) / 180 * PI) * dmoving;
 
 	// up(jump)
-	dmoving = get_moving(Key::Space);
 	if (m_is_keys_pressed[Key::Space])
 	{
 		if (m_flying) {
-			m_dpos.y += dmoving*3;
+			dmoving = get_moving(Key::Space);
+
+			m_dpos.y += dmoving;
 			m_on_ground = false;
 		}
 		else {
-			if (m_on_ground) {
-				m_direction_speed[Key::Space] = 2*COORDS_IN_BLOCK;
+			if (m_is_in_water) {
+				m_dpos.y += 3.f *dtime;
 				m_on_ground = false;
-			}
-			else {
-				m_direction_speed[Key::Space] = 0;
+				m_direction_speed[Key::Space] = 2.5f;
+			} else if (m_on_ground) {
+				m_on_ground = false;
+				m_direction_speed[Key::Space] = 3.f;
 			}
 		}
 	}
@@ -128,56 +129,56 @@ void Player::update(float dtime)
 	if (m_is_keys_pressed[Key::LShift])
 	{
 		if (m_flying) {
-			m_dpos.y -= dmoving*3;
+			m_dpos.y -= dmoving;
 		}
 	}
 
 
-	if (m_flying) {
-		m_on_ground = false;
-
-		m_pos.x += m_dpos.x;
-		collision(m_dpos.x, 0, 0);
-
-		m_pos.y += m_dpos.y;
-		collision(0, m_dpos.y, 0);
-
-		m_pos.z += m_dpos.z;
-		collision(0, 0, m_dpos.z);
-
-		m_dpos.x = m_dpos.y = m_dpos.z = 0;
-	} else {
+	if (!m_flying) {
 		if (!m_on_ground) {
 			float s = m_direction_speed[Key::Space];
 
-			m_direction_speed[Key::Space] -= 9.8f * dtime;
+			if (s < 50.f) {
 
-			s += m_direction_speed[Key::Space];
-			s *= dtime / 2;
 
-			m_dpos.y += s*5;
+				if (m_is_in_water) {
+					if (!m_is_keys_pressed[Key::Space]) {
+						m_dpos.y -= 3.f*dtime;
+					}
+				}
+				else {
+					m_direction_speed[Key::Space] -= 9.8f * dtime;
+
+					s += m_direction_speed[Key::Space];
+					s *= dtime / 2.f;
+					m_dpos.y += s*3;
+				}
+			}
+			else {
+				m_direction_speed[Key::Space] = 50.f;
+				m_dpos.y += m_direction_speed[Key::Space];
+			}
+
 		}
-
-		m_on_ground = false; //reset
-
-		m_pos.x += m_dpos.x;
-		collision(m_dpos.x, 0, 0);
-
-		m_pos.y += m_dpos.y;
-		collision(0, m_dpos.y, 0);
-
-		m_pos.z += m_dpos.z;
-		collision(0, 0, m_dpos.z);
-
-		m_dpos.x = m_dpos.z = 0;
 	}
+	m_on_ground = false; //reset
+
+	m_pos.x += m_dpos.x;
+	collision(m_dpos.x, 0, 0);
+
+	m_pos.y += m_dpos.y;
+	collision(0, m_dpos.y, 0);
+
+	m_pos.z += m_dpos.z;
+	collision(0, 0, m_dpos.z);
 }
 
 void Player::flight_on()
 {
 	if (!m_flying) {
 		m_flying = true;
-		m_speed *= 1.5;
+		m_speed = 2.5f*DEFAULT_PLAYER_SPEED;
+		m_direction_speed[Key::Space] = 0;
 	}
 }
 
@@ -186,6 +187,7 @@ void Player::flight_off()
 	if (m_flying) {
 		m_flying = false;
 		m_speed = DEFAULT_PLAYER_SPEED;
+		m_direction_speed[Key::Space] = 0;
 	}
 }
 
@@ -198,17 +200,21 @@ void Player::put_block()
 
 	prev_x = x; prev_y = y; prev_z = z;
 	bool able_create = false;
-	for (int distance = 0; distance < 50 * COORDS_IN_BLOCK; ++distance) {
-		
-		x += -sinf(m_camera_angle.x / 180 * PI)/10.F;
-		y += tanf(m_camera_angle.y / 180 * PI)/10.F;
-		z += -cosf(m_camera_angle.x / 180 * PI)/10.F;
 
-		if (!m_map->is_air(Map::coord2block_coord(x), Map::coord2block_coord(y), Map::coord2block_coord(z))) {
+	auto dist = [&]()->float { return std::sqrt((m_pos.x - x)*(m_pos.x - x) + (m_pos.y - y)*(m_pos.y - y) + (m_pos.y - y)*(m_pos.y - y)); };
+	while (dist() < 6.f) {
+		
+		x += -sinf(m_camera_angle.x / 180 * PI)/20.F;
+		y += tanf(m_camera_angle.y / 180 * PI)/20.F;
+		z += -cosf(m_camera_angle.x / 180 * PI)/20.F;
+
+		if (!m_map->is_air(Map::coord2block_coord(x), Map::coord2block_coord(y), Map::coord2block_coord(z)) &&
+			!m_map->is_water(Map::coord2block_coord(x), Map::coord2block_coord(y), Map::coord2block_coord(z))
+			) {
 			// is we in this block
-			for (int matrix_x = Map::coord2block_coord(m_pos.x - m_size.x); matrix_x <(m_pos.x + m_size.x) / COORDS_IN_BLOCK; matrix_x++) {
-				for (int matrix_y = Map::coord2block_coord(m_pos.y - m_size.y); matrix_y < (m_pos.y + m_size.y) / COORDS_IN_BLOCK; matrix_y++) {
-					for (int matrix_z = Map::coord2block_coord(m_pos.z - m_size.z); matrix_z < (m_pos.z + m_size.z) / COORDS_IN_BLOCK; matrix_z++) {
+			for (int matrix_x = Map::coord2block_coord(m_pos.x - m_size.x); matrix_x < m_pos.x + m_size.x; matrix_x++) {
+				for (int matrix_y = Map::coord2block_coord(m_pos.y - m_size.y); matrix_y < m_pos.y + m_size.y; matrix_y++) {
+					for (int matrix_z = Map::coord2block_coord(m_pos.z - m_size.z); matrix_z < m_pos.z + m_size.z; matrix_z++) {
 						int x = Map::coord2block_coord(prev_x);
 						int y = Map::coord2block_coord(prev_y);
 						int z = Map::coord2block_coord(prev_z);
@@ -239,19 +245,26 @@ void Player::put_block()
 void Player::delete_block()
 {
 	float x = m_pos.x,
-		y = m_pos.y+ m_size.y*0.8f,
+		y = m_pos.y + 0.8f,
 		z = m_pos.z;
 
 	bool able_create = false;
-	for (int distance = 0; distance < 50 * COORDS_IN_BLOCK; ++distance) {
-		x += -sinf(m_camera_angle.x / 180 * PI)/10.F;
-		y += tanf(m_camera_angle.y / 180 * PI)/10.F;
-		z += -cosf(m_camera_angle.x / 180 * PI)/10.F;
+
+	auto dist = [&]()->float { return std::sqrt((m_pos.x - x)*(m_pos.x - x) + (m_pos.y - y)*(m_pos.y - y) + (m_pos.y - y)*(m_pos.y - y)); };
+	while (dist() < 6) {
+		x += -sinf(m_camera_angle.x / 180 * PI)/20.F;
+		y += tanf(m_camera_angle.y / 180 * PI)/20.F;
+		z += -cosf(m_camera_angle.x / 180 * PI)/20.F;
 		if (!m_map->is_air(
 			Map::coord2block_coord(x),
 			Map::coord2block_coord(y),
-			Map::coord2block_coord(z)
-		)) {
+			Map::coord2block_coord(z)) &&
+
+			!m_map->is_water(
+			Map::coord2block_coord(x),
+			Map::coord2block_coord(y),
+			Map::coord2block_coord(z))
+			) {
 			m_map->delete_block(
 				Map::coord2block_coord(x),
 				Map::coord2block_coord(y),
@@ -264,21 +277,35 @@ void Player::delete_block()
 
 void Player::collision(float dx, float dy, float dz)
 {
+	m_is_in_water = false;
+
 	//for  blocks in player's area
-	for (int x = Map::coord2block_coord(m_pos.x - m_size.x); x < (m_pos.x + m_size.x) / COORDS_IN_BLOCK; x++) {
-		for (int y = Map::coord2block_coord(m_pos.y - m_size.y); y < (m_pos.y + m_size.y) / COORDS_IN_BLOCK; y++) {
-			for (int z = Map::coord2block_coord(m_pos.z - m_size.z); z < (m_pos.z + m_size.z) / COORDS_IN_BLOCK; z++) {
+	for (int y = Map::coord2block_coord(m_pos.y - m_size.y); y < m_pos.y + m_size.y; y++) {
+
+		for (int x = Map::coord2block_coord(m_pos.x - m_size.x); x < m_pos.x + m_size.x; x++) {
+			for (int z = Map::coord2block_coord(m_pos.z - m_size.z); z < m_pos.z + m_size.z; z++) {
 				if (m_map->is_solid(x, y, z)) { //if collided with block
-					if (dx > 0)  m_pos.x = x * COORDS_IN_BLOCK - m_size.x;
-					if (dx < 0)  m_pos.x = x * COORDS_IN_BLOCK + COORDS_IN_BLOCK + m_size.x;
-					if (dy > 0)  m_pos.y = y * COORDS_IN_BLOCK - m_size.y;
+					if (dx > 0)  m_pos.x = x - m_size.x;
+					if (dx < 0)  m_pos.x = x + 1 + m_size.x;
+					if (dy > 0)  m_pos.y = y - m_size.y;
 					if (dy < 0) {
-						m_pos.y = y * COORDS_IN_BLOCK + COORDS_IN_BLOCK + m_size.y;
+						m_pos.y = y + 1 + m_size.y;
 						m_on_ground = true;
-						m_dpos.y = 0;
+						m_direction_speed[Key::Space] = 0.f;
 					}
-					if (dz > 0)  m_pos.z = z * COORDS_IN_BLOCK  - m_size.z;
-					if (dz < 0)  m_pos.z = z * COORDS_IN_BLOCK + COORDS_IN_BLOCK  + m_size.z;
+					if (dz > 0)  m_pos.z = z - m_size.z;
+					if (dz < 0)  m_pos.z = z + 1  + m_size.z;
+				}
+
+				if (m_map->is_water(x, y, z)) {
+					m_is_in_water = true;
+				}
+
+				if (m_map->is_water(x, y, z)) {
+					m_is_under_water = true;
+				}
+				else {
+					m_is_under_water = false;
 				}
 			}
 		}
@@ -371,9 +398,9 @@ void Player::mouse_input(sf::Event& e)
 	static int curr_block_index = 0;
 
 	if (e.type == sf::Event::MouseButtonReleased) {
-		if (e.key.code == sf::Mouse::Left) {
+		if (e.key.code == sf::Mouse::Right) {
 			put_block();
-		} else if (e.key.code == sf::Mouse::Right) {
+		} else if (e.key.code == sf::Mouse::Left) {
 			delete_block();
 		}
 	} else if (e.type == sf::Event::MouseWheelMoved) {
